+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 ### Features
 - `IPv6Addr` now returns its canonical string representation. This TypeScript implementation is ported over from Rust's implementation (`impl Display for std::net::Ipv6Addr`).
 
+### Internal
+- `Ipv4Addr.tryNew()` now has more extensive unit tests.
+
 ## 0.7.6
 
 ### Features